Instead of forcing (require 'package-name) to run…
…check instead to see if the package exists in the first place, since
that's what you want to do anyway, isn't it? You just want to check if
the package exists, and then run the code if it does.

In any case, we're only looking at the two machines I use on a regular
basis, and the common files. In any case, only changes that seem to need
to be done appear to be with the friend desktop. Will need to test out
the MASB machine and see if there are any problems.