Documentation Changes
I discovered here-documents. So then I proceeded to the following:

- Make the installation completely executable by itself: all you need to
do, I think, is copy the entire thing and make it executable and it
should run.
- Incorporate not only the installation of pygame's documentation, but
python's as well.

What I could do next is make it a little smarter and check to see if
python's and pygame's info files are installed. I could. See how.