Documentation cleanup.
It was going to be another minor documentation cleanup, before I
realized: you know what, I'm kind of done with dealing with the
checklists. They don't even work properly with github.

So I basically removed most of them. I've keept the Eee-PC setup, mostly
because I don't know if I want to use that machine in the future (I
think I will; but not as vanilla debian; must consider).