Partial fix to package-list
Some things we've discovered so far: that package-install won't take a
list of strings, it needs to be a list of symbols. So, by using
make-symbol, we convert the list of strings into something that
technically package-install should be able to use.

Except that it isn't. What's happening now instead is that
package-install keeps insisting that 'dired+' and 'org-plus-contrib'
doesn't exist. Running package-install directly works, both as
eval-expression and on the lisp interaction mode, but otherwise it
doesn't work on the init file. Most vexing.

In any case, this is a partial solution. This branch is, for now, broken.