Used local var with lexical binding for prettify-symbols-mode
Thanks to phils and Drew from StackOverflow, I've managed to use local
variables instead of scribbling a global variable, which is nice. It's a
bit iffy, though, since we're using lexical let instead of vanilla let,
which is a bit of a letdown, and I'm still using lambdas. Baby steps,
though.