Revamped tariqk-common, found and killed small bug in friend
So basically, tariqk-common was getting a little unwieldy, so I
rearranged the notes to better organize everything. Also, I've added
el-doc mode, pushed that and the prettify-symbol-mode heading under the
same section, AND upgraded PSM to use the same list of Emacs
Lisp-related modes at the beginning of this heading.

And I killed the source of that weird error message I got from friend.
It was a matter of quoting, really. Tested the offending sexp on
Windows, should work great on Linux.

Stuff looks great now. Loving it! <3