Cleaned up package management in friend
Did the whole table thing with friend's configuration. In the process,
discovered that rather than using mapc to perform the mapping, just call
starter-kit-install-if-needed with *all* the packages. This way all I
need to do is, at most, two calls to that function. Sadly it's not so
straightforward, so what happened was that we ended up learning how to
use apply. Another new thing learned today, yay.