Minor changes
Really, the problem I find with these org-files is that bugs sit
lurking for the longest time before they get resolved, because they
don't get run often enough.

Off the top of my head:
- Changed some named list to tables, because list functionality is
funky.
- Changed `org-twbs` to `MELPA` instead of MELPA-stable, because it
was throwing me errors based on an error that wasn't fixed in the
stable branch.