Added paradox-package manager
Added it because it does async updates and upgrades, which is very
much appreciated. Apparently Emacs 25 has async updates, but paradox
does both updates and upgrades, so I'm going with that.

Also, using the MELPA-stable branch. There's apparently a
`paradox-require` mechanism, but meh, I'm already using this
functionality, why mess with it?